[OSD-8943] Move Handler to types to allow import from other packages
