+++ docs/types.md
@@ -1,0 +1,5 @@
+# Types
+
+## Handler
+
+Defines `Handler` type with additional `Addroute` method which adds the handler's route the to the mux router.
